Added new policies store files
