+++ src/app/state-management/effects/policies.effects.ts
@@ -1,115 +1,107 @@
 import { Injectable } from "@angular/core";
-import { Observable, of } from "rxjs";
-import { Actions, Effect, ofType } from "@ngrx/effects";
-import { Action, select, Store } from "@ngrx/store";
-import { PolicyApi } from "shared/sdk/services";
+import { Actions, createEffect, ofType } from "@ngrx/effects";
+import { PolicyApi, Policy } from "shared/sdk";
+import { Store, select } from "@ngrx/store";
 import {
-  FETCH_POLICIES,
-  FetchPoliciesAction,
-  FetchPoliciesCompleteAction,
-  FetchPoliciesFailedAction,
-  SUBMIT_POLICY,
-  SubmitPolicyAction,
-  SubmitPolicyCompleteAction,
-  SubmitPolicyFailedAction,
-  CHANGE_PAGE,
-  FetchCountPolicies,
-  FailedPoliciesAction,
-  FETCH_EDITABLE_POLICIES,
-  FetchEditablePoliciesComplete,
-  FetchEditablePolicies,
-  SORT_BY_COLUMN
-} from "../actions/policies.actions";
-import { getQueryParams, getPolicies } from "../selectors/policies.selectors";
+  getQueryParams,
+  getPolicies
+} from "state-management/selectors/policies.selectors";
+import * as fromActions from "state-management/actions/policies.actions";
 import {
-  catchError,
-  map,
-  mergeMap,
   switchMap,
-  withLatestFrom
+  withLatestFrom,
+  map,
+  catchError,
+  mergeMap
 } from "rxjs/operators";
-import { Policy } from "state-management/models";
-import * as userSelectors from "state-management/selectors/users.selectors";
+import { of } from "rxjs";
+import { getProfile } from "state-management/selectors/users.selectors";
 
 @Injectable()
-export class PoliciesEffects {
-  @Effect()
-  submitPolicy$: Observable<Action> = this.actions$.pipe(
-    ofType(SUBMIT_POLICY),
-    map((action: SubmitPolicyAction) => action),
-    switchMap(action =>
-      this.policyApi
-        .updatewhere(action.ownerList.join(), action.policyAttributes)
-        .pipe(
-          mergeMap((data: any) => [
-            new SubmitPolicyCompleteAction(data.submissionResponse),
-            new FetchPoliciesAction()
-          ]),
-          catchError(err => of(new SubmitPolicyFailedAction(err)))
-        )
-    )
-  );
-
-  @Effect({ dispatch: false })
+export class PolicyEffects {
   private queryParams$ = this.store.pipe(select(getQueryParams));
-
-  @Effect()
-  private fetchPolicies$: Observable<Action> = this.actions$.pipe(
-    ofType(FETCH_POLICIES, CHANGE_PAGE, SORT_BY_COLUMN),
-    withLatestFrom(this.queryParams$),
-    map(([action, params]) => params),
-    switchMap(({ limits }) =>
-      this.policyApi.find(limits).pipe(
-        mergeMap(policies => [
-          new FetchPoliciesCompleteAction(policies as Policy[]),
-          new FetchEditablePolicies()
-        ]),
-        catchError(err => of(new FetchPoliciesFailedAction()))
-      )
-    )
-  );
-
-  @Effect()
-  FetchCountPolicies$ = this.actions$.pipe(
-    ofType(FETCH_POLICIES),
-    switchMap(action =>
-      this.policyApi.count().pipe(
-        map(({ count }) => new FetchCountPolicies(count)),
-        catchError(err => of(new FailedPoliciesAction(err)))
-      )
-    )
-  );
-
-  @Effect({ dispatch: false })
-  private userProfile$ = this.store.pipe(select(userSelectors.getProfile));
-
-  @Effect({ dispatch: false })
   private policies$ = this.store.pipe(select(getPolicies));
+  private userProfile$ = this.store.pipe(select(getProfile));
 
-  @Effect()
-  FetchEditablePolicies$ = this.actions$.pipe(
-    ofType(FETCH_EDITABLE_POLICIES),
-    withLatestFrom(this.userProfile$),
-    withLatestFrom(this.policies$),
-    map(([[action, profile], allPolicies]) => {
-      const editablePolicies = [];
-      if (!profile) {
-        // allow functional users
-        return new FetchEditablePoliciesComplete(allPolicies);
-      }
-      const email = profile.email.toLowerCase();
-      allPolicies.forEach(pol => {
-        if (pol.manager.indexOf(email) !== -1) {
-          editablePolicies.push(pol);
+  fetchPolicies$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(
+        fromActions.fetchPoliciesAction,
+        fromActions.changePageAction,
+        fromActions.sortByColumnAction
+      ),
+      withLatestFrom(this.queryParams$),
+      map(([action, params]) => params),
+      switchMap(params =>
+        this.policyApi.find(params).pipe(
+          mergeMap((policies: Policy[]) => [
+            fromActions.fetchPoliciesCompleteAction({ policies }),
+            fromActions.fetchEditablePoliciesAction()
+          ]),
+          catchError(() => of(fromActions.fetchPoliciesFailedAction()))
+        )
+      )
+    )
+  );
+
+  fetchCount$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(fromActions.fetchPoliciesAction),
+      switchMap(() =>
+        this.policyApi.count().pipe(
+          map(({ count }) => fromActions.fetchCountCompleteAction({ count })),
+          catchError(() => of(fromActions.fetchCountFailedAction()))
+        )
+      )
+    )
+  );
+
+  fetchEditablePolicies$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(fromActions.fetchEditablePoliciesAction),
+      withLatestFrom(this.userProfile$, this.policies$),
+      map(([action, profile, allPolicies]) => {
+        const editablePolicies = [];
+        if (!profile) {
+          return fromActions.fetchEditablePoliciesCompleteAction({
+            policies: allPolicies
+          });
+        } else {
+          const email = profile.email.toLowerCase();
+          allPolicies.forEach(policy => {
+            if (policy.manager.indexOf(email) !== -1) {
+              editablePolicies.push(policy);
+            }
+          });
+          return fromActions.fetchEditablePoliciesCompleteAction({
+            policies: editablePolicies
+          });
         }
-      });
-      return new FetchEditablePoliciesComplete(editablePolicies);
-    })
+      }),
+      catchError(() => of(fromActions.fetchEditablePoliciesFailedAction()))
+    )
+  );
+
+  submitPolicy$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(fromActions.submitPolicyAction),
+      switchMap(({ ownerList, policy }) =>
+        this.policyApi.updatewhere(ownerList.join(), policy).pipe(
+          mergeMap(({ submissionResponse }) => [
+            fromActions.submitPolicyCompleteAction({
+              policy: submissionResponse
+            }),
+            fromActions.fetchPoliciesAction()
+          ]),
+          catchError(() => of(fromActions.submitPolicyFailedAction()))
+        )
+      )
+    )
   );
 
   constructor(
     private actions$: Actions,
-    private store: Store<any>,
-    private policyApi: PolicyApi
+    private policyApi: PolicyApi,
+    private store: Store<Policy>
   ) {}
 }
